commander: move level calibration out of accelerometer_calibration.cpp
