+++ src/modules/commander/CMakeLists.txt
@@ -46,6 +46,7 @@ px4_add_module(
 		commander_helper.cpp
 		esc_calibration.cpp
 		gyro_calibration.cpp
+		level_calibration.cpp
 		mag_calibration.cpp
 		rc_calibration.cpp
 		state_machine_helper.cpp
