+++ src/modules/commander/level_calibration.h
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2013 PX4 Development Team. All rights reserved.
+ *   Copyright (c) 2013-2020 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -31,20 +31,9 @@
  *
  ****************************************************************************/
 
-/**
- * @file accelerometer_calibration.h
- *
- * Definition of accelerometer calibration.
- *
- * @author Anton Babushkin <anton.babushkin@me.com>
- */
-
-#ifndef ACCELEROMETER_CALIBRATION_H_
-#define ACCELEROMETER_CALIBRATION_H_
+#pragma once
 
 #include <stdint.h>
 #include <uORB/uORB.h>
 
-int do_accel_calibration(orb_advert_t *mavlink_log_pub);
-
-#endif /* ACCELEROMETER_CALIBRATION_H_ */
+int do_level_calibration(orb_advert_t *mavlink_log_pub);
